requirements.txt
joblib==1.5.2
numpy==2.2.6
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
polars-1.39.3